Address missed comments
